Tutorial: Providing images for <img>
I tried to setup an endpoint for providing binary data for <img> tags.
This is what I found out.

I need some help with the following point.
- Return 404 when the requested id is not present in table.

If we could trigger raw output via a config parameter from within the
procedure, we could provide a generic file endpoint. Setting
`raw-media-types= */*` is probably not a good idea. Is this worth a
feature request?

Addresses #233.